requirements.txt
Flask==1.1.2
flask-cors==1.10.3
grpcio==1.32.0
h5py==2.10.0
image==1.5.33
imutils==0.5.4
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.2
scikit-image==0.17.2
scipy==1.5.4
tqdm==4.57.0
memoization==0.3.2
func-timeout==4.3.5
py-sudoku==1.0.1